Fix download file naming
Newer webkitgtk seems to be giving the entire URI to the script making
filenames like "http---...". Also, fix the regex in sed so as to not
replace valid characters (such as '.', '-', and ':'). The '\W' might
also be a GNU-ism (it's from perl and the non-"word" characters seem to
be defined as "not allowed in a perl identifier" which isn't exactly
what we want).